refactor: Refactor tokens layout a bit
